Admin dashboard template functions complete

+++ controllers/api/rm-template-routes.js
@@ -4,12 +4,16 @@ const Template = require("../../models/Template");
 const Category = require("../../models/Category");
 const { withAuth, withAdminAuth } = require("../../utils/auth");
 const router = require("express").Router();
-const { formatTemplateListItems } = require("../../utils/html-utils");
+const {
+  formatTemplateListItems,
+  sanitizeHTML,
+} = require("../../utils/html-utils");
 const {
   getTemplateById,
   getTemplateEditData,
   getAllCategories,
 } = require("../../utils/model-utils");
+const { addHTMLTags } = require("../../utils/markdown-utils");
 
 // route to create a new Template
 router.post("/", withAdminAuth, async (req, res) => {
@@ -48,11 +52,12 @@ router.post("/", withAdminAuth, async (req, res) => {
     }
 
     const newTemplate = await Template.create({
-      markdown,
       title,
-      administrator_id,
-      category_id,
       description,
+      category_id,
+      markdown,
+      html: "",
+      administrator_id,
       public: publicTemplate,
     });
 
@@ -146,33 +151,32 @@ router.get("/new", withAdminAuth, async (req, res) => {
     res.render("template-edit", templateData);
   } catch (error) {
     console.log(error);
-    res.status(404).render("error-404", { message: error });
+    res.status(404).json( { message: error });
   }
 });
 
 // route to edit a template via handlebars
 router.get("/edit", withAdminAuth, async (req, res) => {
-  console.log(req.params);
   try {
     let template_id = req.query.id;
     let returnPath = req.query.path;
 
     if (!template_id) {
-      res.status(404).render("error-404", { message: "Template not found" });
+      res.status(404).json( { message:"Template not found" });
       return;
     }
     if (isNaN(template_id)) {
-      res.status(404).render("error-404", { message: "Template not found" });
+      res.status(404).json( { message:"Template not found" });
       return;
     }
     template_id = parseInt(template_id);
     if (template_id < 1) {
-      res.status(404).render("error-404", { message: "Template not found" });
+      res.status(404).json( { message:"Template not found" });
       return;
     }
     const templateData = await getTemplateEditData(template_id);
     if (!templateData) {
-      res.status(404).render("error-404", { message: "Template not found" });
+      res.status(404).json( { message:"Template not found" });
       return;
     }
     templateData.administrator_id = req.session.user_id;
@@ -185,25 +189,50 @@ router.get("/edit", withAdminAuth, async (req, res) => {
     res.render("template-edit", templateData);
   } catch (error) {
     console.log(error);
-    res.status(404).render("error-404", { message: error });
+    res.status(404).json( { message: error });
   }
 });
-
-router.get("/:id", withAuth, async (req, res) => {
+// gets a template, adds inline css to the html, returns the html
+router.get("/formatted/:id", withAuth, async (req, res) => {
   let template_id = req.params.id;
   if (!template_id) {
-    res.status(404).render("error-404", { message: "Template not found" });
+    res.status(404).json({ message: "Template not found" });
     return;
   }
   if (isNaN(template_id)) {
-    res.status(404).render("error-404", { message: "Template not found" });
+    res.status(404).json({ message: "Template not found" });
     return;
   }
   template_id = parseInt(template_id);
 
   const templateData = await getTemplateById(template_id);
   if (!templateData) {
-    res.status(404).render("error-404", { message: "Template not found" });
+    res.status(404).json({ message: "Template not found" });
+    return;
+  }
+  let html = templateData.html;
+  html = sanitizeHTML(html);
+  html = await addHTMLTags(html);
+  html = `<div class="markdown">${html}</div>`;
+
+  res.status(200).send(html);
+});
+
+router.get("/:id", withAuth, async (req, res) => {
+  let template_id = req.params.id;
+  if (!template_id) {
+    res.status(404).json( { message:"Template not found" });
+    return;
+  }
+  if (isNaN(template_id)) {
+    res.status(404).json( { message:"Template not found" });
+    return;
+  }
+  template_id = parseInt(template_id);
+
+  const templateData = await getTemplateById(template_id);
+  if (!templateData) {
+    res.status(404).json( { message:"Template not found" });
     return;
   }
   res.status(200).json(templateData);
@@ -211,7 +240,6 @@ router.get("/:id", withAuth, async (req, res) => {
 
 // route a for a search on the templates model using either an id or text search
 router.post("/search", withAuth, async (req, res) => {
-  console.log(req.body);
   let { id, searchTerm, searchMarkdown, returnFormat } = req.body;
 
   // validate the search criteria
@@ -341,8 +369,15 @@ router.post("/search", withAuth, async (req, res) => {
 });
 
 // delete a template
-router.delete("/admin/:id", withAdminAuth, async (req, res) => {
-  const template_id = req.params.id;
+router.delete("/delete/:id", withAdminAuth, async (req, res) => {
+  let template_id = req.params.id;
+  if(!template_id) {
+    return res.status(404).json({ message: "Template not found" });
+  }
+  if(isNaN(template_id)) {
+    return res.status(404).json({ message: "Template not found" });
+  }
+  template_id = parseInt(template_id);
 
   try {
     const deletedTemplate = await Template.destroy({
@@ -364,10 +399,15 @@ router.delete("/admin/:id", withAdminAuth, async (req, res) => {
 });
 
 // delete a comment
-// ! TO DO
 router.delete("/comment/", withAdminAuth, async (req, res) => {
-  const template_id = req.body.id;
-
+  let template_id = req.body.id;
+  if(!template_id) {
+    return res.status(404).json({ message: "Template not found" });
+  }
+  if(isNaN(template_id)) {
+    return res.status(404).json({ message: "Template not found" });
+  }
+  template_id = parseInt(template_id);
   try {
     const deletedTemplate = await Template.destroy({
       where: { id: template_id },
